Fix the five engines that crashed on entry
All five died inside their *_engine_start. Three separate bugs, found by
booting each scene directly rather than navigating to it.

* bmp_font_obj_print_text was an auto_stub returning 0, and every OBJ-font
  path in bitmap_font.c goes through it — callers dereference the result on
  the next line. Translated from asm/code_08003980/asm_080049f0.s (204 lines
  of Thumb); its six callees were already in C. The GBA version hardcodes
  offsets into PrintedTextAnim that assume a 4-byte pointer, so the
  translation uses the struct members and sizeof instead. This alone fixed
  four of the five.

* sprite_id_set_data took its generic argument as u32, losing the top half of
  the s16*[2] that SPRITE_ACT_SET_ORIGIN_XY passes; and the handler for that
  case stepped to the second element with (u32 *)arg + 1, reading it from the
  middle of the first.

* bg_anim_update_script shifted the event number right by 28. That is correct
  for the GBA's packed u32 script word, but the PC BgMapAnim keeps the event
  in its own field. Every line therefore read as event 0, GOTO never fired,
  and the reader ran off the end of the script into garbage.

Also adds RTPC_SCENE=<name>, which boots straight into any of the 146 scenes.
Navigating menus to reach a given engine is not practical to do for all of
them, and most had never been run at all; tools/gen_scene_table.py builds the
name table from the Scene definitions in the build.

Crashes across the 55 rhythm-game scenes: 5 -> 0.

+++ src/lib_0804ca80.h
@@ -258,7 +258,7 @@ extern u32  sprite_get_cel_dimensions(u16 *cel, u32 requestedDataType);
 extern intptr_t sprite_get_data(struct SpriteHandler *, s16 id, u32 requestedDataType);
 extern void sprite_set_callback_cel(struct SpriteHandler *, s16 id, s8 cel);
 extern void sprite_run_callback_every_cel(struct SpriteHandler *, s16 id);
-extern void sprite_id_set_data(struct SpriteHandler *, u16 memID, u32 targetDataType, u32 arg);
+extern void sprite_id_set_data(struct SpriteHandler *, u16 memID, u32 targetDataType, uintptr_t arg);
 extern u16  func_0804e0a0(struct SpriteHandler *); // Increment Unknown Counter
 extern void sprite_handler_set_mem_id(struct SpriteHandler *, u16 memID);
 extern u16  sprite_handler_get_mem_id(struct SpriteHandler *);
